+++ src/game/player/elements/scene/Scene.tsx
@@ -64,6 +64,10 @@ export default function Scene(
             });
         }
     });
+    
+    useEffect(() => {
+        state.events.emit(GameState.EventTypes["event:state.onRender"]);
+    }, []);
 
     return (
         <div className={clsx(className, "w-full h-full absolute")}>